requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
langchain-community
requests
python-dotenv
duckduckgo-search
fuzzywuzzy
pyautogui
pynput
python-Levenshtein
pywin32
livekit
ultralytics
opencv-python
easyocr
pypdf
pyperclip

pyttsx3
langdetect
googletrans==4.0.0-rc1
PySide6
psutil
qtawesome
qt-material